requirements.txt
# Add your Python and frontend requirements here

# LLM & Agents
openai==1.50.2
langchain==0.1.17
llama-index==0.10.32
langchain-core
langchain-community
langgraph==0.0.38
tiktoken

# Vector DB
pinecone-client>=3.2.2,<4.0.0

# Backend
fastapi==0.110.1
uvicorn==0.29.0
psutil

# Frontend
streamlit==1.35.0
altair==5.3.0
plotly==5.21.0

# Data
pandas==2.0.3
duckdb==0.10.1

# Tools
python-dotenv==1.0.1
httpx==0.27.0
fpdf
python-multipart
PyPDF2
# For advanced file loaders
beautifulsoup4
python-pptx
nbformat
# langchain_community loaders cover markdown, html, pptx, notebook

pytest
